requirements.txt
jupyterlab
mlflow
python-dotenv